ml-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
transformers>=4.35.2
torch>=2.9.0
sentencepiece>=0.1.99
numpy>=1.24.3
pandas>=2.1.3
vaderSentiment>=3.3.2
nltk>=3.8.1
requests>=2.31.0
pydantic-settings>=2.1.0
# spacy>=3.7.2  # Optional: requires C++ compiler on Windows


